+++ TenantScoreSheet/Models/PointsSummary.cs
@@ -7,13 +7,13 @@
         public int? ApplicantId { get; set; }
         public int? TenantSNo { get; set; }
 
-        public int? TotalPoints { get; set; }
+        public double? TotalPoints { get; set; }
         public bool? FinalApproval { get; set; }
-        public int? TotalDeposit { get; set; }
-        public int? DepositToHoldpaid { get; set; }
-        public int? PetDeposit { get; set; }
-        public int? AdditionalDeposit { get; set; }
-        public int? BalanceDepositDue { get; set; }
+        public double? TotalDeposit { get; set; }
+        public double? DepositToHoldpaid { get; set; }
+        public double? PetDeposit { get; set; }
+        public double? AdditionalDeposit { get; set; }
+        public double? BalanceDepositDue { get; set; }
         public string? CreatedBy { get; set; } = string.Empty!;
         public DateTime? CreatedDate { get; set; }
         public string? ModifiedBy { get; set; } = string.Empty!;
